Nucleus 2G: Rename files for better conformance to naming conventions
